requirements.txt
yfinance
pandas
PyYAML
python-dotenv
requests
numpy
